fix bugs and add feature
#1 add random initial SN generation
#2 fix uO_X_random filed buf memcpy addr error
#3 add sid support for static IP_ID case
#4 fix sn sent in NBO
#5 fix payload buffer pointer error
#6 add align as 1 byte for udp and ipv4 header struct
#7 add ntoh and htno wrapper functions
#8 fix incorrect preamble of packet type (SVDL types)
#9 fix src port from uint32 to uint16

+++ app/udpip_builder.h
@@ -8,7 +8,7 @@ extern bool build_udp_ip_headers(uint8_t *const pkt_buf_start,
                                  const uint16_t payload_len,
                                  const uint32_t src_addr,
                                  const uint32_t dst_addr,
-                                 const uint32_t src_port,
-                                 const uint32_t dst_port,
+                                 const uint16_t src_port,
+                                 const uint16_t dst_port,
                                  const uint16_t id);
 #endif /* ROHC_COMP_APP_UDP_IP_BUILDER_H */
